+++ changelog.md
@@ -5,6 +5,7 @@ Add `/rc start` command which either shows the usage pop-up or starts the addon 
 ## Bugfixes
 
 - *Fixed potential nil error (Classic#46, CurseClassic#170).*
+- *Fixed issue preventing the addon from addon items to the trade window.*
 
 # v3.6.0